WIP on database refactor
